requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
boto3==1.26.74
botocore==1.29.124
Django==4.1.7
django-cors-headers==3.14.0
django-phonenumber-field==7.1.0
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
joblib==1.2.0
numpy==1.24.3
pandas==1.5.1
phonenumbers==8.13.11
Pillow==9.4.0
psycopg2==2.9.6
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2023.3
urllib3==1.26.15